SEQNG-860 Support the offset column
