+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/sequence/steps/StepSettings.scala
@@ -19,7 +19,8 @@ import seqexec.model.enumerations
 import seqexec.web.client.components.SeqexecStyles
 import seqexec.web.client.model.Pages
 import seqexec.web.client.model.lenses._
-import seqexec.web.client.model.ModelOps._
+import seqexec.web.client.model.StepItems._
+import seqexec.web.client.model.Formatting._
 import seqexec.web.client.semanticui.elements.label.Label
 import seqexec.web.client.semanticui.elements.icon.Icon._
 import seqexec.web.client.semanticui.Size
